Improve discount calculator
